directions/serializers.py
from rest_framework import serializers
from .models import *
class TrendSerializer(serializers.ModelSerializer):
    class Meta:
        model = Trend
        fields = ['id', 'title', 'imageActive','imageInactive',]

class DirectionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Direction
        fields = ['id', 'title',]
