requirements.txt
Django==2.2.2
django-admin-sortable2==0.7.3
django-import-export==1.2.0
django-nested-admin==3.2.3
django-suit==0.2.26
django-tabbed-admin==1.0.4
python-monkey-business==1.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Pillow==6.1.0
djangorestframework==3.9.4
django-cleanup==4.0.0
django-rest-swagger==2.2.0
django-cors-headers==3.1.0
django-image-cropping==1.2.0
django-ckeditor==5.7.1
psycopg2==2.8.4